# backend/api-docs/api-docs.yaml
swagger: "2.0"
info:
  description: "This is the documentation for the use of the Youdemy API. If you are interested in obtaining information from Youdemy to process yourself you are at the right place. Take a look at the existing endpoints. Remember to add /api/ before any of these calls"
  version: "1.0.0"
  title: "Youdemy"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8443"
basePath: "/"
...
schemes:
- "https"
- "http"
paths:
  /courses:
    get:
      tags:
      - "courses"
      summary: "Get all courses"
      description: ""
      operationId: "addcourses"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/courses"
        "400":
          description: "Invalid status value"
      security:
      - coursesorders_auth:
        - "write:coursess"
        - "read:coursess"
    put:
      tags:
      - "courses"
      summary: "Update an existing course"
      description: ""
      operationId: "updatecourses"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "courses object that needs to be added to the orders"
        required: true
        schema:
          $ref: "#/definitions/courses"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "courses not found"
        "405":
          description: "Validation exception"
      security:
      - coursesorders_auth:
        - "write:coursess"
        - "read:coursess"
    post:
      tags:
      - "courses"
      summary: "Create a new course"
      description: ""
      operationId: "createCourse"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "courses object that needs to be added to the orders"
        required: true
        schema:
          $ref: "#/definitions/courses"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "courses not found"
        "405":
          description: "Validation exception"
      security:
      - coursesorders_auth:
        - "write:coursess"
        - "read:coursess"
  /courses/{coursesId}:
    get:
      tags:
      - "courses"
      summary: "Find courses by ID"
      description: "Returns a single courses"
      operationId: "getcoursesById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "coursesId"
        in: "path"
        description: "ID of courses to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/courses"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "courses not found"
      security:
      - api_key: []
...
securityDefinitions:
  coursesorders_auth:
    type: "oauth2"
    authorizationUrl: "http://coursesorders.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:coursess: "modify coursess in your account"
      read:coursess: "read your coursess"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Category:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Category"
...
  Tag:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Tag"
  courses:
    type: "object"
    required:
    - "name"
    - "photoUrls"
    properties:
      id:
        type: "integer"
        format: "int64"
      category:
        $ref: "#/definitions/Category"
      name:
        type: "string"
        example: "doggie"
      photoUrls:
        type: "array"
        xml:
          name: "photoUrl"
          wrapped: true
        items:
          type: "string"
      tags:
        type: "array"
        xml:
          name: "tag"
          wrapped: true
        items:
          $ref: "#/definitions/Tag"
      status:
        type: "string"
        description: "courses status in the orders"
        enum:
        - "available"
        - "pending"
        - "sold"
    xml:
      name: "courses"
...